fix: bind no1 native build and frozen artifacts
